Remove observation.py (moved into rpi.py)
