+++ rpi.py
@@ -1,17 +1,44 @@
+from time import sleep
 import cv2
 import RPi.GPIO as GPIO
-import observation
-import preprocessor
 
+import preprocessor
+import camera
+
+# SERVO SETTINGS
 SERVO_PIN = 22
 FREQUENCY = 50  # in hz
 
-reverse = False
-i = 0
+# Field of View for Camera
+START = 60
+END = 120
+INC = 20
+
+def set_pos(pwm, pos):
+    dc_range = (5, 10)
+    dc = dc_range[0] + (pos / 180) * (dc_range[1] - dc_range[0])  # Duty cycle
+    pwm.ChangeDutyCycle(dc)
+    sleep(0.1)
+    pwm.ChangeDutyCycle(0)
+
+def capture_view(camera_idx, pwm, reverse = False) -> list:
+    images = []
+    start = START if not reverse else END
+    end = END + 1 if not reverse else START - 1
+    inc = INC if not reverse else -INC
+    for pos in range(start, end, inc):
+        set_pos(pwm, pos)
+        images.append(camera.get_frame(camera_idx))
+        sleep(0.75)
+    if reverse:
+        images.reverse()
+    return images
+
+
+reverse = False  # Optimization - please ignore
 
 def get_frame(camera_idx):
     global reverse
-    global i  # TODO: Remove i
     GPIO.setmode(GPIO.BOARD)
     GPIO.setup(SERVO_PIN, GPIO.OUT)
 
@@ -22,11 +49,7 @@ def get_frame(camera_idx):
     images = observation.capture_view(camera_idx, pwm, reverse)
     # for i in range(len(images)):
     #     cv2.imwrite(f"img{i}.png", images[i])
-    print("Combining images and writing to file")
-    cv2.imwrite(f"result{i}.png", preprocessor.combine_images(images))  # Return the image instead of saving to file
-    i += 1
-    # TODO: Send result to pipeline
-    reverse = not reverse
-
     pwm.stop()
     GPIO.cleanup()
+    print(f"Combining {len(images)} images")
+    return preprocessor.combine_images(images)
